# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo API Documentation",
        "description": "Application Programming Interface Of Demo",
        "contact": {
            "email": "smile@tyyi.net"
        },
        "license": {
            "name": "MIT",
            "url": "http://access_is_not_allowed_by_third_party"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Machine"
        },
        {
            "url": "http://api.demo.com:8000",
            "description": "Demo Api Server"
        }
    ],
    "paths": {
        "/api/v1/cities": {
            "get": {
                "tags": [
                    "Geo"
                ],
                "summary": "Get list of cities",
                "description": "Returns list of cities",
                "operationId": "getCities",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            },
            "post": {
                "tags": [
                    "Geo"
                ],
                "summary": "Stores a new city",
                "description": "Stores record in the database",
                "operationId": "insertCity",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/City"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/cities/{id}": {
            "get": {
                "tags": [
                    "Geo"
                ],
                "summary": "Returns a single city",
                "description": "Retreives record from database",
                "operationId": "getCity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the city which is asked for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Geo"
                ],
                "summary": "Updates a single city",
                "description": "Updates a record in database",
                "operationId": "122930c2c25829e631432afaa727a3ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/City"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Geo"
                ],
                "summary": "Delete Existing City",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteCity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "City": {
                "title": "City",
                "description": "City Model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "string",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the new city",
                        "type": "string",
                        "example": "همدان"
                    },
                    "province_id": {
                        "title": "Province ID",
                        "description": "ID of the province containing this city",
                        "type": "integer",
                        "format": "string",
                        "example": 1
                    },
                    "latitude": {
                        "title": "Latitude",
                        "description": "Geographical latitude of this city",
                        "type": "string",
                        "example": "37° 55’ 4.008’’ N"
                    },
                    "longitude": {
                        "title": "Longitude",
                        "description": "Geographical longitude of this city",
                        "type": "string",
                        "example": "46° 7’ 28.085’’ E"
                    },
                    "phi": {
                        "title": "Phi",
                        "description": "Geographical φ(d) of this city",
                        "type": "string",
                        "example": "37.917781829834"
                    },
                    "lambda": {
                        "title": "Lambda",
                        "description": "Geographical  λ(d) of this city",
                        "type": "string",
                        "example": "46.1244697570801"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "City"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {jwt})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}